requirements.txt
aiosmtpd==1.4.6
alembic==1.13.1
atpublic==4.1.0
attrs==23.2.0
BareNecessities==0.2.8
blinker==1.8.2
cffi==1.16.0
click==8.1.7
clickhouse-driver==0.2.8
colorama==0.4.6
cryptography==42.0.7
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mail==2.1.0
Mako==1.3.5
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
numpy==2.0.0
packaging==24.0
pandas==2.2.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
Werkzeug==3.0.3
WTForms==3.1.2
